UefiPayloadPkg: Add support for Firmware Volume Block Protocol
Use coreboot's SMMSTORE v2 to implement a FVB protocol.
Add a library for SMMStore to be used in PEI and DXE.

The PEI component is important as the VariableStore needs to be initialized
for the RuntimeVariableDxe, and the Pcd must be pointed to the mmaped SMMStore
in memory.

The DXE component provides runtime services and takes care of virtual to physical
mapping the communication buffers between SMM and OS.

MD: Adapted from 9e commit c9fdcb7a

Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>