UefiPayloadPkg/BlSupportDxe: Use EfiSmbiosProtocol to install tables
The default EfiSmbiosProtocol operates on an empty SMBIOS table.
As the SMBIOS tables are provided by the bootloader, install the
SMBIOS tables using the EfiSmbiosProtocol.
For this introduce a new DXE called BlSupportDxeSmbios which is only
dispatched when the gEfiSmbiosProtocolGuid is present.

This fixes the settings menu not showing any hardware information, instead
only "0 MB RAM" was displayed.

Tests showed that the OS can still see the SMBIOS tables.

Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>