MdeModulePkg/UsbBusDxe: Reset USB port on device error
USB device enumeration will skip a hot reset for a port if a reset change status is detected; this can happen after a soft reset on a host controller as this also hot resets the individual ports.

An attached device might not respond if the port reset change bits are set, resulting from a hot reset
that transitions to a warm reset and doesn't complete it successfully.

To mitigate this, reset a port if a querying returns an error.

Cc: Hao A Wu <hao.a.wu@intel.com>
Cc: Ray Ni <ray.ni@intel.com>
Signed-off-by: Sean Rhodes <sean@starlabs.systems>