MdeModulePkg/UsbBusDxe: Adjust USB Enumeration flow
When edk2 requests the first device descriptor for the bMaxPacketSize0, only
the first 8 bytes within the device descriptor are requested. Particular USB
 drives will fail to provide the complete descriptor after this request; this
results in the USB drive not being enumerated.

Adjust the enumeration flow as follows:
1. Reset the port in the event of a status change error
2. Port Stabilization Debounce
3. First Port Reset
4. First Device Descriptor Request for bMaxPacketSize0.
5. Second Port Reset
6. Additional Reset
7. Set the USB address
8. Second Device Descriptor Request

This patch also adjusts the requirements for the MaxPacketLength, to match
what is seen on effected USB drives.

With this new flow, 99% of USB drives tested will be enumerated successfully.

Signed-off-by: Sean Rhodes <sean@starlabs.systems>
Signed-off-by: Matt DeVillier <matt.devillier@gmail.com>
Change-Id: I91fefdc681f7bce4f4dfb64275e2ab62c9c1ec3a