MdeModulePkg/XhciDxe/XhciReg: Add handing for full-speed USB2 devices
On some platforms (SKL/KBL), the Protocol Speed ID Value (PSIV) indices
are incorrectly shared between Protocol Speed ID dwords in the extended
capabilites registers for USB2 and USB3. This leads to USB2/Full Speed
devices incorrectly being identified as USB3/Super Speed. To correct
this, check the USB2 extended caps before the USB3 caps, and exit if a
match is found in the PSI dwords for a Full Speed device, so that the
USB3 caps are not checked.

Signed-off-by: Matt DeVillier <matt.devillier@gmail.com>
Change-Id: I5bcf32105ce85fda95b4ba98a5e420e8f522374c