Fix Network stack not being loaded (#26)
Due to the changes in the commit  
4c4ceb2c, the Network Stack requires the use of a new Random Number Generator (RNG).
The UefiPayloadPkg has not been patched on default EDK2, since network implementation is not default in it. Therefore, the solution was to apply something that was done in another platforms in EDK2 adding the RngDxe and the Hash2DxeCrypto drivers (commits 3e722403 and e10d8323).